# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OverUnder API", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/kyc/session": {
      "post": {
        "tags": ["kyc"],
        "summary": "Kyc Session",
        "description": "Create or update KYC session for the authenticated user.\n\nThis endpoint does NOT store government IDs or documents.\nIt only stores status enum and jurisdiction.",
        "operationId": "kyc_session_api_v1_kyc_session_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KycSessionRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KycSessionResponse" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/kyc/status": {
      "get": {
        "tags": ["kyc"],
        "summary": "Kyc Status",
        "description": "Get KYC status for the authenticated user.",
        "operationId": "kyc_status_api_v1_kyc_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KycSessionResponse" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/kyc/check": {
      "post": {
        "tags": ["kyc"],
        "summary": "Kyc Check",
        "description": "Check if KYC is required based on notional amount and jurisdiction.\n\nGate logic:\n- notional ≥ threshold (default $500/day) OR restricted jurisdiction → require KYC\n- Operator JWT cannot bypass fiat KYC\n- Returns allowed=True if KYC status is 'pass' or 'not_required'",
        "operationId": "kyc_check_api_v1_kyc_check_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KycCheckRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KycCheckResponse" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/api/v1/oracle/attest": {
      "post": {
        "tags": ["oracle"],
        "summary": "Attest",
        "description": "Persist one agent report (the oracle job sends its operator JWT).",
        "operationId": "attest_api_v1_oracle_attest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttestIn" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/oracle/resolved": {
      "post": {
        "tags": ["oracle"],
        "summary": "Mark Resolved",
        "operationId": "mark_resolved_api_v1_oracle_resolved_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResolvedIn" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/oracle/{market_id}/status": {
      "get": {
        "tags": ["oracle"],
        "summary": "Status",
        "operationId": "status_api_v1_oracle__market_id__status_get",
        "parameters": [
          { "name": "market_id", "in": "path", "required": true, "schema": { "type": "string", "title": "Market Id" } }
        ],
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/api/v1/oracle/vote": {
      "post": {
        "tags": ["oracle"],
        "summary": "Vote",
        "description": "One vote per holder per market. voter = the caller; weight = their CTF YES+NO balance.",
        "operationId": "vote_api_v1_oracle_vote_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/VoteIn" } } }
        },
        "responses": {
          "200": { "description": "Successful Response", "content": { "application/json": { "schema": {} } } },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttestIn": {
        "properties": {
          "conditionId": { "type": "string", "maxLength": 66, "title": "Conditionid" },
          "agent": { "type": "string", "maxLength": 42, "title": "Agent" },
          "outcome": { "type": "integer", "maximum": 255.0, "minimum": 0.0, "title": "Outcome" },
          "evidenceHash": { "type": "string", "maxLength": 66, "title": "Evidencehash" },
          "summary": { "type": "string", "title": "Summary", "default": "" },
          "evidenceJson": { "type": "string", "title": "Evidencejson", "default": "[]" },
          "signature": { "type": "string", "title": "Signature", "default": "" }
        },
        "type": "object",
        "required": ["conditionId", "agent", "outcome", "evidenceHash"],
        "title": "AttestIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KycCheckRequest": {
        "properties": { "notional_usdc": { "type": "number", "exclusiveMinimum": 0.0, "title": "Notional Usdc" } },
        "type": "object",
        "required": ["notional_usdc"],
        "title": "KycCheckRequest"
      },
      "KycCheckResponse": {
        "properties": {
          "allowed": { "type": "boolean", "title": "Allowed" },
          "reason": { "type": "string", "title": "Reason", "default": "" },
          "kyc_status": { "type": "string", "title": "Kyc Status", "default": "" }
        },
        "type": "object",
        "required": ["allowed"],
        "title": "KycCheckResponse"
      },
      "KycSessionRequest": {
        "properties": { "jurisdiction": { "type": "string", "title": "Jurisdiction", "default": "" } },
        "type": "object",
        "title": "KycSessionRequest"
      },
      "KycSessionResponse": {
        "properties": {
          "status": { "type": "string", "title": "Status" },
          "jurisdiction": { "type": "string", "title": "Jurisdiction" },
          "updated_at": { "type": "string", "title": "Updated At" }
        },
        "type": "object",
        "required": ["status", "jurisdiction", "updated_at"],
        "title": "KycSessionResponse"
      },
...
      "ResolvedIn": {
        "properties": {
          "conditionId": { "type": "string", "title": "Conditionid" },
          "outcome": { "type": "integer", "title": "Outcome" }
        },
        "type": "object",
        "required": ["conditionId", "outcome"],
        "title": "ResolvedIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "VoteIn": {
        "properties": {
          "conditionId": { "type": "string", "maxLength": 66, "title": "Conditionid" },
          "outcome": { "type": "integer", "maximum": 1.0, "minimum": 0.0, "title": "Outcome" },
          "voter": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Voter" },
          "weight": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Weight" }
        },
        "type": "object",
        "required": ["conditionId", "outcome"],
        "title": "VoteIn"
      }
    }
  }
}